login/register

+++ app/Http/Controllers/Home/PostController.php
@@ -47,11 +47,17 @@ class PostController extends Controller
      * @param     Request
      * @return    [type]
      */
-    public function store(Request $request)
+    public function store()
     {
         //校验
+        $this->validate(request(),[
+            'title'=>'required|string|min:5|max:100',
+            'content'=>'required|string|min:10',
+        ]);
+        $user_id=\Auth::id();
+        $params=array_merge(request(['title','content']),compact('user_id'));
         //入库
-        Post::create($request->all());
+        Post::create($params);
         return redirect('/');
     }
     /**
@@ -71,10 +77,21 @@ class PostController extends Controller
      * @DateTime  2018-05-08
      * @return    [type]
      */
-    public function update()
+    public function update(Post $post)
     {
-        dd(request());
-        Post::update();
+        //验证
+        $this->validate(request(),[
+            'title'=>'required|string|min:5|max:100',
+            'content'=>'required|string|min:10',
+        ]);
+        //权限验证
+        $this->authorize('update',$post);
+        //入库
+        $post->title=request('title');
+        $post->content=request('content');
+        $post->save();
+        //渲染
+        return redirect('/post/{$post->id}');
     }
     /**
      * 删除文章
@@ -84,6 +101,8 @@ class PostController extends Controller
      */
     public function delete(Post $post)
     {
+        //权限验证
+        $this->authorize('delete',$post);
         $post->delete();
         return redirect('/');
     }
